chapter_004/src/main/java/ru/job4j/stream/School.java
package ru.job4j.stream;

import ru.job4j.pojo.Student;

import java.util.List;
import java.util.Map;
import java.util.function.Predicate;
import java.util.stream.Collectors;

public class School {
    public static List<Student> collect(List<Student> students, Predicate<Student> predict) {
        return students.stream()
                .filter(predict)
                .collect(Collectors.toList());
    }

    public static Map<String, Student> collectToMap(List<Student> students) {
        return students.stream()
                .collect(Collectors.toMap(
                        e -> e.getSurname(),
                        e -> e
                ));
    }
}
